Refactor result transformer
