+++ quesma/quesma/transformations.go
@@ -4,6 +4,7 @@ package quesma
 
 import (
 	"quesma/model"
+	"strings"
 )
 
 type TransformationPipeline struct {
@@ -15,4 +16,22 @@ func (o *TransformationPipeline) Transform(queries []*model.Query) ([]*model.Que
 		queries, _ = transformer.Transform(queries)
 	}
 	return queries, nil
+}
+
+type replaceColumNamesWithFieldNames struct {
+}
+
+func (t *replaceColumNamesWithFieldNames) Transform(result [][]model.QueryResultRow) ([][]model.QueryResultRow, error) {
+
+	const doubleColons = "::"
+	const dot = "."
+
+	for _, rows := range result {
+		for i, row := range rows {
+			for j := range row.Cols {
+				rows[i].Cols[j].ColName = strings.ReplaceAll(row.Cols[j].ColName, doubleColons, dot)
+			}
+		}
+	}
+	return result, nil
 }
